pyproject.toml
[project]
name = "backend-budget-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.19.0",
    "fastapi>=0.104.1",
    "greenlet>=3.0.1",
    "pandas>=2.1.4",
    "pydantic>=2.5.1",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "uvicorn>=0.24.0",
]
